delete Evidence and Create User

+++ controllers/evidence.js
@@ -53,13 +53,63 @@ evidenveRouter.get('/:id',async(req,res,next) => {
 
 })
 
-evidenveRouter.delete('/:id',(req,res,next) => {
-  const id = req.params.id
-  Evidenve.findByIdAndDelete(id)
-    .then(() => {
-      res.status(204).end()
-    })
-    .catch(error => next(error))
+evidenveRouter.delete('/:id',async (req,res,next) => {
+  try {
+    //Authorizaction
+    const token = getTokenFrom(req)
+    const decodedToken = jwt.verify(token,process.env.SECRET)
+    if(!token||!decodedToken){
+      return res.status(401).json({ error: 'token missing or invalid' })
+    }else{
+      const rolID = decodedToken.rol
+      const user = decodedToken.id
+      const rol = await Rol.findById(rolID)
+      if(!rol){
+        return res.status(401).json({ error: 'rol missing or invalid' })
+      }else if(rol.name!==ROL_ADMIN){
+        console.log('rol',rol.name)
+        return res.status(401).json({ error: 'unauthorized rol' })
+      }
+      //end-authorization
+      const id = req.params.id
+      const evidenceDelete = await Evidenve.findByIdAndDelete(id)
+      //console.log(evidenceDelete)
+      const subindicatorBD = await Subindicator.findById(evidenceDelete.subIndicatorID)
+        .populate({
+          path:'evidences'
+        })
+        .populate({
+          path:'typeID',
+          populate:{
+            path:'characteristics'
+          }
+        })
+      //console.log(subindicatorBD)
+      //const x = subindicatorBD.evidences.findIndex(evi => evi.equals(evidenceDelete._id))
+      //console.log(x)
+      console.log('1:',subindicatorBD.evidences.length)
+      //subindicatorBD.evidences.splice(x,1)
+      const subindicatorUpdated = autoQualifySubindicator(subindicatorBD)
+      subindicatorUpdated.lastUpdate=new Date()
+      subindicatorUpdated.lastUpdateBy = user
+      console.log('2:',subindicatorUpdated.evidences.length)
+      await Subindicator.findByIdAndUpdate(subindicatorUpdated.id,subindicatorUpdated,{ new:true })
+
+      //console.log(subindicatorUpdated)
+      const indicatorBD = await IndicatorInstance.findById(subindicatorUpdated.indicadorID)
+        .populate({
+          path: 'subindicators',
+        })
+      //console.log(indicatorBD)
+      const inddicatorUpdated = autoQualificationIndicator(indicatorBD)
+      inddicatorUpdated.lastUpdate=new Date()
+      inddicatorUpdated.lastUpdateBy = user
+      await IndicatorInstance.findByIdAndUpdate(inddicatorUpdated.id,inddicatorUpdated,{ new:true })
+      return res.json({ message:'evidencia eliminada con exito' })
+    }
+  } catch (error) {
+    next(error)
+  }
 })
 evidenveRouter.post('/',async (req,res,next) => {
   try {
@@ -184,6 +234,126 @@ evidenveRouter.put('/:id',(req,res,next) => {
     .catch(error => next(error))
 })
 module.exports = evidenveRouter
+const autoQualificationIndicator = (indicator) => {
+  const arraySubindicators = indicator.subindicators
+
+  // Variables auxiliares para contar los subindicadores por calificación
+  let count1 = 0
+  let count2 = 0
+  let count3 = 0
+
+  // Iterar sobre los subindicadores y contar las calificaciones
+  arraySubindicators.forEach((subindicator) => {
+    if (subindicator.qualification === 1) {
+      count1++
+    } else if (subindicator.qualification === 2) {
+      count2++
+    } else if (subindicator.qualification === 3) {
+      count3++
+    }
+  })
+
+  // Calcular el número total de subindicadores evaluados
+  const numberEvaluated = count1 + count2 + count3
+  let qualification=0
+  // Actualizar la calificación del indicador según las condiciones
+  //si tengo un rojo todo esta en rojo
+  if (count1 > 0) {
+    indicator.qualification = 1
+    qualification=1
+    console.log('1')
+  //si tengo un amarillo y numero de ealuador es mayor a la mitad de suubindicadores estas en amarillo
+  } else if (count2 > 0 && numberEvaluated > arraySubindicators.length / 2) {
+    indicator.qualification = 2
+    qualification=2
+    console.log('2')
+  //si tengo todas en verde esta en verde el indicador
+  } else if (count3 === arraySubindicators.length && count3>4) {
+    indicator.qualification = 3
+    qualification=3
+    console.log('3')
+  //si tengo mas de la mitad en verde estas en amarillo
+  }else if (count3>arraySubindicators.length/2){
+    indicator.qualification = 2
+    qualification=2
+    console.log('2')
+  //si numero de evualos es mayor 0 estas en rojo
+  } else if (numberEvaluated > 0) {
+    indicator.qualification = 1
+    qualification=1
+    console.log('1')
+    //console.log('check')
+  //si no ningun subindicador evaluado estas en gris
+  } else {
+    indicator.qualification = 0
+    qualification=0
+  }
+  console.log(qualification)
+  return indicator
+}
+const autoQualifySubindicator = (subindcatorBD) => {
+  const arrayCharacteristics = subindcatorBD.typeID.characteristics
+  //red
+  let existEvidence = []
+  //yellow
+  let existEvidenceCritic = []
+  const arrayEvidences = subindcatorBD.evidences
+  let qualifySubindicator = 0
+  let scoreSubindicator = 0
+  arrayCharacteristics.forEach(characteristic => {
+    const total = characteristic.score
+    scoreSubindicator+=total
+    const founded = arrayEvidences.filter(evidence => evidence.characteristicID.equals(characteristic._id))
+    //comprobar si existe evidencia para c/u caracteristica
+    if(founded.length>0){
+      existEvidence.push(true)
+      let sum = 0
+      founded.forEach(evidence => {
+        sum+=evidence.qualification
+      })
+      const qualify = sum/total
+      qualifySubindicator+=qualify
+    }else{
+      existEvidence.push(false)
+    }
+    //comprobar si falta evidencia para una caracteristica critica
+    if((characteristic.tier>1 && founded.length===0) ){
+      existEvidenceCritic.push(false)
+    }else{
+      //si la caracteristica no es critica o no hay evidencia
+      existEvidenceCritic.push(true)
+    }
+  })
+  console.log('score',scoreSubindicator)
+  console.log('qualify',qualifySubindicator)
+  subindcatorBD.score=qualifySubindicator
+  subindcatorBD.totalScore=scoreSubindicator
+
+  const count = existEvidence.reduce((acc,curr) => {
+    if(curr){
+      acc.trueCount++
+    }else{
+      acc.falseCount++
+    }
+    return acc
+  },{ trueCount:0,falseCount:0 })
+  const total = arrayCharacteristics.length
+  const percent = count.trueCount/total
+  //console.log(existEvidence)
+  //Si existen todas las evidencias = verde
+  if(!existEvidence.includes(false)){
+    subindcatorBD.qualification=3
+  //Si hay mas del 50% de evidencias = yellow
+  }else if(count.trueCount>count.falseCount){
+    subindcatorBD.qualification=2
+    //Si falta una evidencia critica || hay mas del 10% de evidencias pero menos del 50% = rojo
+  }else if(existEvidenceCritic.includes(false)|| (percent>0.1 && percent<=0.5)){
+    subindcatorBD.qualification=1
+  }else {
+    subindcatorBD.qualification=0
+  }
+  return subindcatorBD
+}
 
 const updateSubindicator = async(evidence) => {
   const subindicatorID =String(evidence.subIndicatorID)
@@ -280,68 +450,7 @@ const updateSubindicator2 = async(evidence) => {
         path:'characteristics'
       }
     })
-  const arrayCharacteristics = subindcatorBD.typeID.characteristics
 
-  //red
-  let existEvidence = []
-  //yellow
-  let existEvidenceCritic = []
-  //subindcatorBD.evidences = subindcatorBD.evidences.concat(evidence)
-  const arrayEvidences = subindcatorBD.evidences
-  let qualifySubindicator = 0
-  let scoreSubindicator = 0
-  arrayCharacteristics.forEach(characteristic => {
-    const total = characteristic.score
-    scoreSubindicator+=total
-    const founded = arrayEvidences.filter(evidence => evidence.characteristicID.equals(characteristic._id))
-    //comprobar si existe evidencia para c/u caracteristica
-    if(founded.length>0){
-      existEvidence.push(true)
-      let sum = 0
-      founded.forEach(evidence => {
-        sum+=evidence.qualification
-      })
-      const qualify = sum/total
-      qualifySubindicator+=qualify
-    }else{
-      existEvidence.push(false)
-    }
-    //comprobar si falta evidencia para una caracteristica critica
-    if((characteristic.tier>1 && founded.length===0) ){
-      existEvidenceCritic.push(false)
-    }else{
-      //si la caracteristica no es critica o no hay evidencia
-      existEvidenceCritic.push(true)
-    }
-  })
-  console.log('score',scoreSubindicator)
-  console.log('qualify',qualifySubindicator)
-  subindcatorBD.score=qualifySubindicator
-  subindcatorBD.totalScore=scoreSubindicator
-
-  const count = existEvidence.reduce((acc,curr) => {
-    if(curr){
-      acc.trueCount++
-    }else{
-      acc.falseCount++
-    }
-    return acc
-  },{ trueCount:0,falseCount:0 })
-  const total = arrayCharacteristics.length
-  const percent = count.trueCount/total
-  //console.log(existEvidence)
-  //Si existen todas las evidencias = verde
-  if(!existEvidence.includes(false)){
-    subindcatorBD.qualification=3
-  //Si hay mas del 50% de evidencias = yellow
-  }else if(count.trueCount>count.falseCount){
-    subindcatorBD.qualification=2
-    //Si falta una evidencia critica || hay mas del 10% de evidencias pero menos del 50% = rojo
-  }else if(existEvidenceCritic.includes(false)|| (percent>0.1 && percent<=0.5)){
-    subindcatorBD.qualification=1
-  }else {
-    subindcatorBD.qualification=0
-  }
   subindcatorBD.lastUpdate = new Date()
   subindcatorBD.lastUpdateBy = evidence.author//el ultimo oque registro evidencia
   const subindicatorUpdate = await Subindicator.findByIdAndUpdate(subindcatorBD.id,subindcatorBD,{ new:true })//hemos actualizado y recalificado el suubindicador
